unify shared elements
