+++ packages/styleguide/src/lib/Foundations/ColorMode/ColorMode.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/blocks';
 
 import { AboutHeader } from '~styleguide/blocks';
 
-import { DarkModeTable, LightModeTable } from '../Theme/tables';
+import { DarkModeTable, LightModeTable } from '../shared/elements';
 import {
   BackgroundExample,
   ColorModeExample,
