+++ packages/styleguide/src/lib/Foundations/Theme/Theme.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/blocks';
 
 import { AboutHeader, DataTable } from '~styleguide/blocks';
 
-import * as TABLES from './tables';
+import * as TABLES from '../shared/elements';
 
 export const parameters = {
   title: 'Theme',
