+++ packages/styleguide/src/lib/Foundations/Typography.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/blocks';
 
 import { ComponentHeader, DataTable } from '~styleguide/blocks';
 
-import * as TABLES from './Theme/tables';
+import * as TABLES from './shared/elements';
 
 export const parameters = {
   title: 'Typography',
