+++ packages/styleguide/src/lib/Foundations/shared/elements.tsx
@@ -1,5 +1,4 @@
-// eslint-disable gamut/import-paths
-import { Box, DataTable } from '@codecademy/gamut';
+import { Anchor, Box, DataTable } from '@codecademy/gamut';
 import {
   Background,
   coreSwatches,
@@ -7,9 +6,12 @@ import {
   theme,
   trueColors,
 } from '@codecademy/gamut-styles';
+// eslint-disable-next-line gamut/import-paths
+import * as ALL_PROPS from '@codecademy/gamut-styles/src/variance/config';
+import kebabCase from 'lodash/kebabCase';
 
 // CASS - fix
-import { Code, ColorScale } from '~styleguide/blocks';
+import { Code, ColorScale, LinkTo } from '~styleguide/blocks';
 
 const PROP_COLUMN = {
   key: 'key',
@@ -396,3 +398,55 @@ export const DarkModeTable = () => (
     <DataTable {...(darkMode as any)} />
   </Background>
 );
+/* eslint-disable gamut/import-paths */
+
+const PROPERTIES_COLUMN = {
+  key: 'properties',
+  name: 'Properties',
+  size: 'xl',
+  render: ({
+    property,
+    properties = [property],
+  }: {
+    property: string;
+    properties: string[];
+  }) =>
+    properties.map((property) => (
+      <Anchor
+        href={`https://developer.mozilla.org/en-US/docs/Web/CSS/${property}`}
+        target="_blank"
+        rel=""
+      >
+        <Code key={property}>{kebabCase(property)}</Code>
+      </Anchor>
+    )),
+};
+
+const SCALE_COLUMN = {
+  key: 'scale',
+  name: 'Scale',
+  size: 'lg',
+  render: ({ scale }: { scale: string }) => (
+    <LinkTo id={`foundations-theme--${kebabCase(scale)}`}>{scale}</LinkTo>
+  ),
+};
+
+const TRANSFORM_COLUMN = {
+  key: 'transform',
+  name: 'Transform',
+  size: 'fill',
+  render: ({ transform }: any) => transform && <Code>{transform?.name}</Code>,
+};
+
+export const defaultColumns = [
+  PROP_COLUMN,
+  PROPERTIES_COLUMN,
+  SCALE_COLUMN,
+  TRANSFORM_COLUMN,
+];
+
+export const getPropRows = (key: keyof typeof ALL_PROPS) =>
+  Object.entries(ALL_PROPS[key]).map(([prop, config]) => ({
+    id: prop,
+    ...config,
+  }));
